+++ TD.Util/Public/Utilities/Convert-ObjectToHashTable.ps1
@@ -1,0 +1,16 @@
+function Convert-ObjectToHashTable([object[]]$Object)
+{
+    foreach ($o in $Object)
+    {
+        $output = @{}
+        $o | Get-Member -MemberType *Property | ForEach-Object {
+            $value = $o.($_.name)
+            if ($value -is [PSCustomObject])
+            {
+                $value = Convert-ObjectToHashTable -Object $value
+            }
+            $output.($_.name) = $value
+        }
+        $output
+    }
+}