+++ docs/release-notes.md
@@ -4,6 +4,20 @@
 ![PowerShell Gallery](https://img.shields.io/powershellgallery/dt/td.util.svg?label=PSGallery%20Downloads&logo=PowerShell&style=flat-square)
 [![Build Status](https://dev.azure.com/tedon/TD.Deploy/_apis/build/status/ehagen.TD.Util?branchName=master)](https://dev.azure.com/tedon/TD.Deploy/_build/latest?definitionId=52&branchName=master)
 
+## [0.1.26]
+
+```plain
+### Added
+- Added Copy-ObjectPropertyValues
+- Added Ado Get-AdoRepositoryPolicy
+- Added Utility functions Convert-ObjectToHashTable, ConvertTo-TitleCase, Copy-ObjectPropertyValues, Get-ArrayItemCount, Get-RandomString and Merge-Objects
+
+### Changed
+
+- Improved saving Json file with SecureStringStorage
+
+```
+
 ## [0.1.25]
 
 ```plain